requirements.txt
tkinter
os
requests
Pillow
io
json
